Modified forms, displayed database in DataGridViews.

+++ HotelBookingProject/Account.cs
@@ -8,7 +8,7 @@ namespace HotelBookingProject
 {
     class Account
     {
-        public int IdAccount { get; set; }
+        public long IdAccount { get; set; }
         public string Username { get; set; }
         public string Password { get; set; }
